Add option to ignore repeated tag markers
When listing tags of a note, this adds the option to only look for tags which contain a single tag marker at the beginning and ignore the others. So if your tag marker is @, and the option is enabled it will display @some_tag as a tag but ignore @@no_tag.